backend/neiqian-agent-main/requirements.txt
fastapi
uvicorn
aiomysql
openai
pydantic
python-dotenv
